Display carousel images (without the animation) that change according to the active style
