+++ src/ProductDetails/index.js
@@ -1,13 +1,13 @@
 import React, {useState, useContext, useEffect} from 'react';
-import { StateContext, DispatchContext } from './../appState/index.js';
-import Info from './productInfo.js';
-import Description from './productDesc.js';
-import Features from './productFeatures.js';
+import { StateContext} from './../appState/index.js';
+import Info from './components/ProductInfo.js';
+import Description from './components/ProductDesc.js';
+import Features from './components/ProductFeatures.js';
 
 
 function ProductDetails() {
   const [state] = useContext(StateContext);
-  const [, dispatch] = useContext(DispatchContext);
+  // const [, dispatch] = useContext(DispatchContext);
   const [activeProduct, setActiveProduct] = useState({});
   const [styles, setStyles] = useState([]);
